# docs/openapi.json
{
  "info": {
    "title": "Building Compliance Platform",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rules/{rule_id}/review": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "action": "approve",
                "notes": "string",
                "reviewer": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        }
      }
    },
...
    "/api/v1/rulesets/validate": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "geometry": {},
                "ruleset_id": 0,
                "ruleset_slug": "string",
                "ruleset_version": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "citations": [],
                  "results": [
                    {
                      "checked": 0,
                      "citation": {
                        "string": null
                      },
                      "passed": true,
                      "rule_id": "string",
                      "target": "string",
                      "title": "string",
                      "violations": []
                    }
                  ],
                  "ruleset": {
                    "authority": "string",
                    "description": "string",
                    "id": 0,
                    "jurisdiction": "string",
                    "name": "string",
                    "slug": "string",
                    "version": 0
                  },
                  "summary": {
                    "checked_entities": 0,
                    "evaluated_rules": 0,
                    "total_rules": 0,
                    "violations": 0
                  }
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Validate a geometry payload against the requested rule pack."
      }
    },
...
  },
...
}